Update tests in JNI example
