+++ Samples/SwiftJavaExtractJNISampleApp/Sources/MySwiftLibrary/FoundationTypes.swift
@@ -2,7 +2,7 @@
 //
 // This source file is part of the Swift.org open source project
 //
-// Copyright (c) 2025 Apple Inc. and the Swift.org project authors
+// Copyright (c) 2026 Apple Inc. and the Swift.org project authors
 // Licensed under Apache License v2.0
 //
 // See LICENSE.txt for license information
@@ -26,4 +26,20 @@ public func compareDates(date1: Date, date2: Date) -> Bool {
 
 public func dateFromSeconds(_ seconds: Double) -> Date {
   Date(timeIntervalSince1970: seconds)
+}
+
+public func echoUUID(_ uuid: UUID) -> UUID {
+  uuid
+}
+
+public func makeUUID() -> UUID {
+  UUID()
+}
+
+public func echoURL(_ url: URL) -> URL {
+  url
+}
+
+public func makeURL() -> URL {
+  URL(string: "http://example.com")!
 }
