--- Samples/SwiftJavaExtractJNISampleApp/Sources/MySwiftLibrary/UUID.swift
@@ -1,38 +1,0 @@
-//===----------------------------------------------------------------------===//
-//
-// This source file is part of the Swift.org open source project
-//
-// Copyright (c) 2025 Apple Inc. and the Swift.org project authors
-// Licensed under Apache License v2.0
-//
-// See LICENSE.txt for license information
-// See CONTRIBUTORS.txt for the list of Swift.org project authors
-//
-// SPDX-License-Identifier: Apache-2.0
-//
-//===----------------------------------------------------------------------===//
-
-import SwiftJava
-
-#if canImport(FoundationEssentials)
-import FoundationEssentials
-#else
-import Foundation
-#endif
-
-public func echoUUID(_ uuid: UUID) -> UUID {
-  uuid
-}
-
-public func makeUUID() -> UUID {
-  UUID()
-}
-
-public func echoURL(_ url: URL) -> URL {
-  url
-}
-
-public func makeURL() -> URL {
-  URL(string: "http://example.com")!
-}
-
